Extract websocket into eventbus module
