+++ antarest/eventbus/main.py
@@ -1,11 +1,15 @@
+from flask import Flask
+
 from antarest.common.config import Config
 from antarest.common.interfaces.eventbus import IEventBus
 from antarest.eventbus.business.local_eventbus import LocalEventBus
 from antarest.eventbus.business.redis_eventbus import RedisEventBus
 from antarest.eventbus.service import EventBusService
+from antarest.eventbus.web import configure_websockets
 
 
 def build_eventbus(
+    application: Flask,
     config: Config,
     autostart: bool = True,
 ) -> IEventBus:
@@ -17,4 +21,6 @@ def build_eventbus(
         else LocalEventBus(),
         autostart,
     )
+
+    configure_websockets(application, eventbus)
     return eventbus
